renamed test class to what it represents.
