+++ src/test/java/org/saucedemo/login/TestProduct.java
@@ -11,16 +11,16 @@ import org.saucedemo.deeplink.DeepLink;
 import org.saucedemo.factories.DriverFactory;
 import org.saucedemo.factories.TestEnvironment;
 import org.saucedemo.screens.ProductScreen;
-import org.saucedemo.test_lifecycle.RunnerExtension;
+import org.saucedemo.testlifecycle.TestLifecycleExtension;
 
 import java.lang.invoke.MethodHandles;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.saucedemo.deeplink.DeepLink.setDeepLinkUrl;
-import static org.saucedemo.test_lifecycle.TestResult.packUp;
+import static org.saucedemo.testlifecycle.TestResult.packUp;
 
-@ExtendWith(RunnerExtension.class)
+@ExtendWith(TestLifecycleExtension.class)
 class TestProduct {
     private static final String CLASS_NAME = MethodHandles.lookup().lookupClass().getSimpleName();
     private static final Config CONFIG = TestEnvironment.getConfig();
